backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pandas
SQLAlchemy
python-dotenv
rapidfuzz
pytest
httpx 